update service and controller in tasks

+++ src/tasks/tasks.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Body, HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Tasks } from './dto-tasks/entity-tasks/tasks.entity';
@@ -21,15 +21,49 @@ export class TasksService {
             return this.taskRepository.save(createTasksDto);
         }
 
-        update(updateTasksDto: UpdateTasksDto, taskId: number){
-            return this.taskRepository.update(taskId, updateTasksDto);
+        async update(updateTasksDto: UpdateTasksDto, taskId: number, @Body('user_id') userId: number){
+            try {
+                if (taskId > 0) {
+                  const task = await this.taskRepository.findOne({where: {id: taskId}});
+                  if (!task) {
+                    throw new HttpException('task Not Found', HttpStatus.BAD_REQUEST);
+                  }
+                  if (task.user_id !== userId) {
+                    throw new HttpException('You don\'t have permission to update this task', HttpStatus.FORBIDDEN);
+                  }
+                  await this.taskRepository.update(taskId, updateTasksDto);                  
+                  return { message: 'task updated Successfully' };
+                }
+                throw new HttpException('Bad Request', HttpStatus.BAD_REQUEST);
+              } catch (err) {
+                throw err;
+              }
+            // return this.taskRepository.update(taskId, updateTasksDto);
         }
 
         getTask(taskId: number){
             return this.taskRepository.findOne({where: {id: taskId}});
         }
 
-        delete(taskId: number){
-            return this.taskRepository.delete(taskId);
+        async delete(taskId: number, @Body('user_id') userId: number){
+
+            try {
+                if (taskId > 0) {
+                    const task = await this.taskRepository.findOne({where: {id: taskId}});
+                    if (!task) {
+                    throw new HttpException('task Not Found', HttpStatus.BAD_REQUEST);
+                  }
+                  if (task.user_id !== userId) {
+                    throw new HttpException('You don\'t have permission to delete this task', HttpStatus.FORBIDDEN);
+                  }
+                  await this.taskRepository.delete(taskId);
+                  return { message: 'task Deleted Successfully' };
+                }
+                throw new HttpException('Bad Request', HttpStatus.BAD_REQUEST);
+              } catch (err) {
+                console.log(err);
+                throw err;
+              }
+
         }
 }
